+++ compiler/src/org.graalvm.compiler.asm.aarch64.test/src/org/graalvm/compiler/asm/aarch64/test/TestProtectedAssembler.java
@@ -171,11 +171,6 @@ class TestProtectedAssembler extends AArch64Assembler {
     }
 
     @Override
-    protected void csinc(int size, Register dst, Register src1, Register src2, ConditionFlag condition) {
-        super.csinc(size, dst, src1, src2, condition);
-    }
-
-    @Override
     protected void fmovFpu2Cpu(int size, Register dst, Register src) {
         super.fmovFpu2Cpu(size, dst, src);
     }