+++ compiler/src/org.graalvm.compiler.lir.amd64/src/org/graalvm/compiler/lir/amd64/AMD64ArrayEqualsOp.java
@@ -319,29 +319,32 @@ public final class AMD64ArrayEqualsOp extends AMD64ComplexVectorOp {
                 // use the 1-byte-1-byte stride variant for the 2-2 and 4-4 cases by simply shifting
                 // the length
                 masm.align(preferredBranchTargetAlignment(crb));
-                masm.bind(variants[AMD64StrideUtil.getDirectStubCallIndex(Stride.S4, Stride.S4)]);
+                masm.bind(variants[StrideUtil.getDirectStubCallIndex(Stride.S4, Stride.S4)]);
                 masm.shll(length, 1);
                 masm.align(preferredBranchTargetAlignment(crb));
-                masm.bind(variants[AMD64StrideUtil.getDirectStubCallIndex(Stride.S2, Stride.S2)]);
+                masm.bind(variants[StrideUtil.getDirectStubCallIndex(Stride.S2, Stride.S2)]);
                 masm.shll(length, 1);
                 masm.align(preferredBranchTargetAlignment(crb));
-                masm.bind(variants[AMD64StrideUtil.getDirectStubCallIndex(Stride.S1, Stride.S1)]);
+                masm.bind(variants[StrideUtil.getDirectStubCallIndex(Stride.S1, Stride.S1)]);
                 emitArrayCompare(crb, masm, Stride.S1, Stride.S1, Stride.S1, result, arrayA, arrayB, mask, length, done, false);
                 masm.jmp(done);
 
                 for (Stride strideA : new Stride[]{Stride.S1, Stride.S2, Stride.S4}) {
                     for (Stride strideB : new Stride[]{Stride.S1, Stride.S2, Stride.S4}) {
-                        if (strideA.log2 <= strideB.log2) {
+                        if (strideA.log2 == strideB.log2 || !withMask() && strideA.log2 < strideB.log2) {
                             continue;
                         }
-                        masm.align(preferredBranchTargetAlignment(crb));
-                        // use the same implementation for e.g. stride 1-2 and 2-1 by swapping the
-                        // arguments in one variant
-                        masm.bind(variants[AMD64StrideUtil.getDirectStubCallIndex(strideB, strideA)]);
-                        masm.movq(tmp, arrayA);
-                        masm.movq(arrayA, arrayB);
-                        masm.movq(arrayB, tmp);
-                        masm.bind(variants[AMD64StrideUtil.getDirectStubCallIndex(strideA, strideB)]);
+                        if (!withMask()) {
+                            masm.align(preferredBranchTargetAlignment(crb));
+                            // use the same implementation for e.g. stride 1-2 and 2-1 by swapping
+                            // the arguments in one variant
+                            masm.bind(variants[StrideUtil.getDirectStubCallIndex(strideB, strideA)]);
+                            masm.movq(tmp, arrayA);
+                            masm.movq(arrayA, arrayB);
+                            masm.movq(arrayB, tmp);
+                        }
+                        masm.align(crb.target.wordSize * 2);
+                        masm.bind(variants[StrideUtil.getDirectStubCallIndex(strideA, strideB)]);
                         emitArrayCompare(crb, masm, strideA, strideB, strideB, result, arrayA, arrayB, mask, length, done, false);
                         masm.jmp(done);
                     }