+++ sdk/src/org.graalvm.polyglot/src/org/graalvm/polyglot/impl/AbstractPolyglotImpl.java
@@ -76,6 +76,7 @@ import org.graalvm.polyglot.Context;
 import org.graalvm.polyglot.Engine;
 import org.graalvm.polyglot.EnvironmentAccess;
 import org.graalvm.polyglot.HostAccess;
+import org.graalvm.polyglot.HostAccess.MutableTargetMapping;
 import org.graalvm.polyglot.HostAccess.TargetMappingPrecedence;
 import org.graalvm.polyglot.Instrument;
 import org.graalvm.polyglot.Language;
@@ -253,6 +254,8 @@ public abstract class AbstractPolyglotImpl {
         public abstract String validatePolyglotAccess(PolyglotAccess access, Set<String> language);
 
         public abstract void engineClosed(Engine engine);
+
+        public abstract MutableTargetMapping[] getMutableTargetMappings(HostAccess access);
 
     }