+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/jni/JNIAutomaticFeature.java
@@ -27,16 +27,19 @@ package com.oracle.svm.hosted.jni;
 import java.util.Collections;
 import java.util.List;
 
+import org.graalvm.nativeimage.Platforms;
 import org.graalvm.nativeimage.hosted.Feature;
+import org.graalvm.nativeimage.impl.InternalPlatform;
 
 import com.oracle.svm.core.SubstrateOptions;
-import com.oracle.svm.core.feature.InternalFeature;
 import com.oracle.svm.core.feature.AutomaticallyRegisteredFeature;
+import com.oracle.svm.core.feature.InternalFeature;
 
 /**
  * Automatically enables {@link JNIFeature} when specific options are set.
  */
 @AutomaticallyRegisteredFeature
+@Platforms(InternalPlatform.NATIVE_ONLY.class)
 public class JNIAutomaticFeature implements InternalFeature {
     @Override
     public boolean isInConfiguration(IsInConfigurationAccess access) {