+++ truffle/src/com.oracle.truffle.host/src/com/oracle/truffle/host/HostLanguage.java
@@ -158,7 +158,10 @@ final class HostLanguage extends TruffleLanguage<HostContext> {
         } else {
             wrapped = value;
         }
-        return HostObject.forObject(wrapped, hostContext);
+        if (wrapped != null) {
+            return HostObject.forObject(wrapped, hostContext);
+        }
+        return null;
     }
 
     @Override